+++ internal/legacy/hcl2shim/flatmap.go
