+++ internal/legacy/helper/schema/field_reader_config_test.go
@@ -11,7 +11,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/opentofu/opentofu/internal/configs/hcl2shim"
+	"github.com/opentofu/opentofu/internal/legacy/hcl2shim"
 	"github.com/opentofu/opentofu/internal/legacy/helper/hashcode"
 	"github.com/opentofu/opentofu/internal/legacy/tofu"
 )
