+++ internal/legacy/helper/schema/resource_timeout.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/mitchellh/copystructure"
-	"github.com/opentofu/opentofu/internal/configs/hcl2shim"
+	"github.com/opentofu/opentofu/internal/legacy/hcl2shim"
 	"github.com/opentofu/opentofu/internal/legacy/tofu"
 )
