+++ internal/legacy/tofu/resource_test.go
@@ -10,11 +10,11 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/opentofu/opentofu/internal/configs/configschema"
+	"github.com/mitchellh/reflectwalk"
 	"github.com/zclconf/go-cty/cty"
 
-	"github.com/mitchellh/reflectwalk"
-	"github.com/opentofu/opentofu/internal/configs/hcl2shim"
+	"github.com/opentofu/opentofu/internal/configs/configschema"
+	"github.com/opentofu/opentofu/internal/legacy/hcl2shim"
 )
 
 func TestResourceConfigGet(t *testing.T) {
