+++ internal/legacy/tofu/state_test.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/davecgh/go-spew/spew"
 	"github.com/opentofu/opentofu/internal/addrs"
-	"github.com/opentofu/opentofu/internal/configs/hcl2shim"
+	"github.com/opentofu/opentofu/internal/legacy/hcl2shim"
 )
 
 func TestStateValidate(t *testing.T) {
